Control.Comonad.Env

Description

The Env comonad (aka the Coreader, Environment, or Product comonad)

A co-Kleisli arrow in the Env comonad is isomorphic to a Kleisli arrow in the reader monad.

(a -> e -> m) ~ (a, e) -> m ~ Env e a -> m

ComonadEnv class

class Comonad w => ComonadEnv e w | w -> e where

Minimal complete definition

ask

Methods

ask :: w a -> e

asks :: ComonadEnv e w => (e -> e') -> w a -> e'

local :: (e -> e') -> EnvT e w a -> EnvT e' w a

Modifies the environment using the specified function.

The Env comonad

type Env e = EnvT e Identity

env :: e -> a -> Env e a

Create an Env using an environment and a value

runEnv :: Env e a -> (e, a)

The EnvT comonad transformer

data EnvT e w a

Constructors

EnvT e (w a)

runEnvT :: EnvT e w a -> (e, w a)

Re-exported modules

module Control.Comonad

module Control.Comonad.Trans.Class