Data.Functor.Composition

Documentation

class Composition o where

We often need to distinguish between various forms of Functor-like composition in Haskell in order to please the type system. This lets us work with these representations uniformly.

Methods

decompose :: o f g x -> f (g x)

compose :: f (g x) -> o f g x